'long' fixer applied
fix_long applied for Configuration.py

double int fixed in src/python

+++ src/python/WMCore/Services/DBS/DBSWriterObjects.py
@@ -292,7 +292,7 @@ def createDBSFiles(fjrFileInfo, jobType = None, apiRef = None):
             for runinfo in fjrFileInfo.runs:
 
                 run = DbsRun(
-                    RunNumber = long(runinfo),
+                    RunNumber = int(runinfo),
                     NumberOfEvents = 0,
                     NumberOfLumiSections = 0,
                     TotalLuminosity = 0,
@@ -310,12 +310,12 @@ def createDBSFiles(fjrFileInfo, jobType = None, apiRef = None):
         for lumiinfo in fjrFileInfo.getLumiSections():
 
             lumi = DbsLumiSection(
-                LumiSectionNumber = long(lumiinfo['LumiSectionNumber']),
+                LumiSectionNumber = int(lumiinfo['LumiSectionNumber']),
                 StartEventNumber = 0,
                 EndEventNumber = 0,
                 LumiStartTime = 0,
                 LumiEndTime = 0,
-                RunNumber = long(lumiinfo['RunNumber']),
+                RunNumber = int(lumiinfo['RunNumber']),
                 )
 
             # Isnt needed, causes monster slowdown
